AfterScan 1.9.8 (#60)
- Moe code handling tooltips to specific module, to allow sharing it with ALT-Scann8. It might become a git submodule later
